# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders",
                "description": "Returns list of orders",
                "operationId": "getOrdersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No orders found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Store a new order",
                "description": "Stores a new order",
                "operationId": "storeOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a specific order",
                "description": "Returns a specific order",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update a specific order",
                "description": "Updates a specific order",
                "operationId": "updateOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete a specific order",
                "description": "Deletes a specific order",
                "operationId": "deleteOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/OrderProducts": {
            "get": {
                "tags": [
                    "OrderProduct"
                ],
                "summary": "Get all OrderProducts",
                "description": "Display a listing of the resource.",
                "operationId": "d0c5d025ea7fff24b93d411d126bc0bc",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderProduct"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "OrderProduct"
                ],
                "summary": "Create a new OrderProduct",
                "description": "Store a newly created resource in storage.",
                "operationId": "d5dfdadb28304c0af03865895b6048d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OrderProduct created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderProduct"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/OrderProducts/{id}": {
            "get": {
                "tags": [
                    "OrderProduct"
                ],
                "summary": "Get a specific OrderProduct by ID",
                "description": "Display the specified resource.",
                "operationId": "a79df01ece44fda779970ad1f84562e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrderProduct ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderProduct"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OrderProduct not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "OrderProduct"
                ],
                "summary": "Update a specific OrderProduct by ID",
                "description": "Update the specified resource in storage.",
                "operationId": "5971c6cd3c22cbe3efc6ebb9e2f82977",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrderProduct ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OrderProduct updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderProduct"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "OrderProduct not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "OrderProduct"
                ],
                "summary": "Delete a specific OrderProduct by ID",
                "description": "Remove the specified resource from storage.",
                "operationId": "fe3e13b64933c6dd3ab7b60e4c9b6b3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrderProduct ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OrderProduct deleted successfully"
                    },
                    "404": {
                        "description": "OrderProduct not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Artisan": {
                "properties": {
                    "id": {
                        "description": "Artisan ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Artisan creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Artisan update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Consumer": {
                "properties": {
                    "id": {
                        "description": "Consumer ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Consumer creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Consumer update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    }
                },
                "type": "object"
            },
            "DeliveryPersonnel": {
                "properties": {
                    "id": {
                        "description": "DeliveryPersonnel ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "DeliveryPersonnel creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "DeliveryPersonnel update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "description": "Order ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderDate": {
                        "description": "Order date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "orderStatus": {
                        "description": "Order status",
                        "type": "string"
                    },
                    "deliveryAddress": {
                        "description": "Delivery address",
                        "type": "string"
                    },
                    "consumer_id": {
                        "description": "ID of the associated consumer",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Order creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Order update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deliveryPersonnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "consumer": {
                        "$ref": "#/components/schemas/Consumer"
                    },
                    "orderProducts": {
                        "description": "Related OrderProducts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderProduct"
                        }
                    }
                },
                "type": "object"
            },
            "OrderProduct": {
                "properties": {
                    "id": {
                        "description": "OrderProduct ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "quantity": {
                        "description": "OrderProduct quantity",
                        "type": "integer"
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "created_at": {
                        "description": "OrderProduct creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "OrderProduct update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "id": {
                        "description": "Product ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string"
                    },
                    "pricePerPiece": {
                        "description": "Product price per piece",
                        "type": "number",
                        "format": "float"
                    },
                    "minOrder": {
                        "description": "Minimum order quantity",
                        "type": "integer"
                    },
                    "type": {
                        "description": "Product type",
                        "type": "string",
                        "enum": [
                            "Sugar",
                            "Salt"
                        ]
                    },
                    "childType": {
                        "description": "Product child type",
                        "type": "string"
                    },
                    "images": {
                        "description": "Product images",
                        "type": "json"
                    },
                    "ratings": {
                        "description": "Related Ratings",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rating"
                        }
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    },
                    "orderProducts": {
                        "description": "Related OrderProducts",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderProduct"
                        }
                    },
                    "created_at": {
                        "description": "Product creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Product update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Rating": {
                "properties": {
                    "id": {
                        "description": "Rating ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "rating": {
                        "description": "Numeric rating value",
                        "type": "integer"
                    },
                    "ratingType": {
                        "description": "Type of rating (DeliveryPersonnel, Product, Artisan)",
                        "type": "string",
                        "enum": [
                            "DeliveryPersonnel",
                            "Product",
                            "Artisan"
                        ]
                    },
                    "created_at": {
                        "description": "Rating creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Rating update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "delivery_personnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "User first name",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "User last name",
                        "type": "string"
                    },
                    "description": {
                        "description": "User description",
                        "type": "string"
                    },
                    "adresse": {
                        "description": "User address",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "User phone number",
                        "type": "string"
                    },
                    "user_type": {
                        "description": "User type",
                        "type": "string",
                        "enum": [
                            "Artisan",
                            "Consumer",
                            "DeliveryPersonnel"
                        ]
                    },
                    "created_at": {
                        "description": "User creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "User update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deliveryPersonnel": {
                        "$ref": "#/components/schemas/DeliveryPersonnel"
                    },
                    "artisan": {
                        "$ref": "#/components/schemas/Artisan"
                    },
                    "consumer": {
                        "$ref": "#/components/schemas/Consumer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}